google_drive: don't add into message for downloaded message

+++ userbot/modules/google_drive.py
@@ -379,13 +379,13 @@ async def download_gdrive(gdrive, service, uri):
                         display_message = current_message
                     except Exception:
                         pass
-        reply += (
+        await gdrive.edit(
             "`[FILE - DOWNLOAD]`\n\n"
             f"`Name   :`\n`{file_name}`\n"
             f"`Size   :` `{humanbytes(file_size)}`\n"
             f"`Path   :` `{file_path}`\n"
             "`Status :` **OK**\n"
-            "`Reason :` Successfully downloaded...\n\n"
+            "`Reason :` Successfully downloaded..."
         )
         msg = await gdrive.respond("`Answer the question in your BOTLOG group`")
     async with gdrive.client.conversation(BOTLOG_CHATID) as conv:
@@ -773,8 +773,11 @@ async def google_drive(gdrive):
                         f"`Reason :` {str(e)}\n\n"
                     )
                     continue
-            await gdrive.respond(reply, link_preview=False)
-            return await gdrive.delete()
+            if reply:
+                await gdrive.respond(reply, link_preview=False)
+                return await gdrive.delete()
+            else:
+                return
         elif re.findall(r'\bhttps?://.*\.\S+', value) or "magnet:?" in value:
             uri = value.split()
         else:
@@ -800,6 +803,8 @@ async def google_drive(gdrive):
             if reply:
                 await gdrive.respond(reply, link_preview=False)
                 return await gdrive.delete()
+            else:
+                return
         if not uri and not gdrive.reply_to_msg_id:
             return await gdrive.edit(
                 "`[VALUE - ERROR]`\n\n"
